requirements.txt
absl-py==1.4.0
altair==4.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
blinker==1.5
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
comm==0.1.4
commonmark==0.9.1
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
Flask==1.1.1
gast==0.2.2
gitdb==4.0.9
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
gunicorn==19.9.0
h5py==2.10.0
idna==3.6
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.8.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.10.1
joblib==1.1.1
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
mock==5.1.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.6.0
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.1
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==3.19.6
pyarrow==6.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.6.2
Pygments==2.14.0
PyJWT==1.7.1
Pympler==1.0.1
pyparsing==3.1.1
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pywinpty==1.1.6
pyzmq==25.1.2
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
semver==2.13.0
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
streamlit==1.10.0
streamlit-option-menu==0.3.12
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.1
traitlets==4.3.3
typing_extensions==4.1.1
tzdata==2024.1
tzlocal==4.2
urllib3==1.26.18
validators==0.20.0
watchdog==2.3.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.6.6
wincertstore==0.2
wrapt==1.16.0
zipp==3.6.0
